+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * [#4156](https://github.com/rubocop-hq/rubocop/issues/4156): Add command line option `--auto-gen-only-exclude`. ([@Ana06][], [@jonas054][])
+* Add new `Naming/ClassVariableName` cop. ([@ylecuyer][])
 
 ### Bug fixes